Separate ner from classification
